make sure that views from prior versions are removed

+++ src/main/java/org/jumpmind/symmetric/db/h2/H2DbDialect.java
@@ -40,20 +40,18 @@ public class H2DbDialect extends AbstractDbDialect implements IDbDialect {
 
     @Override
     protected boolean doesTriggerExistOnPlatform(String catalogName, String schema, String tableName, String triggerName) {
-        // This is for upgrade to get rid of view.
-        jdbcTemplate.update(String.format("DROP VIEW IF EXISTS %s_VIEW", triggerName));
+        if (jdbcTemplate.queryForInt("select count(*) from INFORMATION_SCHEMA.VIEWS WHERE TABLE_NAME = ?",
+                new Object[] { String.format("%s_VIEW", triggerName) }) > 0) {
+            // This is for upgrade to get rid of view.
+            jdbcTemplate.update(String.format("DROP VIEW IF EXISTS %s_VIEW", triggerName));
+        }
         boolean exists = jdbcTemplate
                 .queryForInt("select count(*) from INFORMATION_SCHEMA.TRIGGERS WHERE TRIGGER_NAME = ?",
                         new Object[] { triggerName }) > 0;
-        if (exists) {
+        if (!exists) {
             exists = jdbcTemplate.queryForInt("select count(*) from INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = ?",
                     new Object[] { String.format("%s_VIEW", triggerName) }) > 0;
-        }
-        if (exists) {
-            // TODO Remove in 2.0.  This is for the move away from using views
-            exists = jdbcTemplate.queryForInt("select count(*) from INFORMATION_SCHEMA.VIEWS WHERE TABLE_NAME = ?",
-                    new Object[] { String.format("%s_VIEW", triggerName) }) == 0;
-        }        
+        }       
         return exists;
     }
